fix: refuse a broken file, and stop capping a default at 32 bits
An unterminated quote swallowed every later row into one cell and
reported nothing, so a definition silently got shorter — the loss
format.md §1 warns about and the error diagnostics.md §1 promises, from
a variant that had never been reachable. Alongside it, the column
verdicts the specification names: a default is as wide as its channel
and out of range past it in either notation, a bit past 64 is out of
range rather than not an integer, favorite writes 0, encode reports a
raw value it truncated, remove_row returns None instead of panicking,
and renumbering counts channels rather than rows.

+++ crates/chdef/src/interchange.rs
@@ -97,7 +97,7 @@ pub struct ChannelJson<'l> {
     pub lsb: f64,
     pub offset: f64,
     pub unit: &'l str,
-    pub default: Option<u32>,
+    pub default: Option<u64>,
     pub format: &'static str,
     /// The bound in its own notation, or empty when unspecified.
     pub min: String,
